CSS transition polyfill initial state improvement
Adds support for CSS transitions when an element doesn't include
transitions in the initial state. This supports the use case where a
mounted component is updated to add CSS transition styles. It should be
re-rendered as an Animated component instead of printing an error about
not having an initial transition state.

The 'useStyleTransition' hook is also moved out of
'createStrictDOMComponent' and into the 'useStyleProps' hook.