Allow text-decoration-* styles to be inherited
Text decorations are not technically inherited, but they have a similar
effect. Text decorations on inline elements apply to all boxes generated
by that element. For block-level elements, the setting applies to all
block-level descendants in the normal flow, as well as an anonymous
inline box that includes all inline children in the normal flow.
However, text decorations are not propagated to absolutely positioned
descendants (this detail can be polyfilled in a future patch.)

Close #195